Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@7869 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ randseed/src/ScreenRand.java
@@ -40,7 +40,7 @@ public class ScreenRand
 				}
 			}
 		}
-		int seed = 0;
+		long seed = 0;
 		for (int i = 0; i < rb.length; i++)
 		{
 			int j = rb[i];
